refactor: extract yaml field-ordering into yaml-field-order.ts
Move the field-ordering family (orderDeviceTypeFields, orderPlacedDeviceFields,
orderRackFields, orderCableFields, orderMetadataFields), KNOWN_TOP_LEVEL_KEYS,
UNSAFE_KEYS, and appendUnknownSections from yaml.ts into a new
utils/yaml-field-order.ts as pure functions, exposed via a single
orderLayoutFields entry point. yaml.ts imports that entry point; both serialize
functions are rewired to it. The schema version gate in parseLayoutObject is
untouched, and yaml.ts keeps all its public exports.
